+++ fiboa_cli/datasets/template.py
@@ -1,0 +1,7 @@
+# from ..util import log
+
+def convert(output_file):
+    """
+    Converts the field boundary datasets to fiboa.
+    """
+    pass